Add Spring MVC 3.2 Servlet based async support
Instead of polling at regular intervals for quotes, the server now
holds up the response until a new quote is available, which results
in fewer requests overall and more optimal latency.

Similarly instead of returning a trade confirmation id and then
polling to obtain the confirmation, the server now holds up the
request and until the confirmation becomes available. Since the
request processing is released immediately, the longer wait should
not affect the ability of the server to process requests while
also eliminating the need for polling